feat: Game event SSE use dependency injection
