+++ Backend/service/delivery/http/v1/sse_handler.go
@@ -1,0 +1,74 @@
+package http
+
+import (
+	"github.com/gin-gonic/gin"
+	"log"
+)
+
+type Event struct {
+	Message       chan string
+	NewClients    chan chan string
+	ClosedClients chan chan string
+	TotalClients  map[chan string]bool
+}
+
+type ClientChan chan string
+
+func NewSSEServer() (event *Event) {
+	event = &Event{
+		Message:       make(chan string),
+		NewClients:    make(chan chan string),
+		ClosedClients: make(chan chan string),
+		TotalClients:  make(map[chan string]bool),
+	}
+
+	go event.listen()
+
+	return
+}
+
+func (stream *Event) listen() {
+	for {
+		select {
+		case client := <-stream.NewClients:
+			stream.TotalClients[client] = true
+			log.Printf("Client added. %d registered clients", len(stream.TotalClients))
+
+		case client := <-stream.ClosedClients:
+			delete(stream.TotalClients, client)
+			close(client)
+			log.Printf("Removed client. %d registered clients", len(stream.TotalClients))
+
+		case eventMsg := <-stream.Message:
+			for clientMessageChan := range stream.TotalClients {
+				clientMessageChan <- eventMsg
+			}
+		}
+	}
+}
+
+func (stream *Event) serveHTTP() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		clientChan := make(ClientChan)
+
+		stream.NewClients <- clientChan
+
+		defer func() {
+			stream.ClosedClients <- clientChan
+		}()
+
+		c.Set("clientChan", clientChan)
+
+		c.Next()
+	}
+}
+
+func HeadersMiddleware() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		c.Writer.Header().Set("Content-Type", "text/event-stream")
+		c.Writer.Header().Set("Cache-Control", "no-cache")
+		c.Writer.Header().Set("Connection", "keep-alive")
+		c.Writer.Header().Set("Transfer-Encoding", "chunked")
+		c.Next()
+	}
+}
